Refactor - Create a unittest file for MessageStore
Taken from unittest_lint.py
